Trekk ut arbeidslistetypar til eiga fil slik at det blir tydelegare at vi vil bli kvitt dei.
